Add CI control files for Travis and AppVeyor.
This commit is not intended to be part of a patch submission, it's just a way
to trigger builds on Travis and/or AppVeyor.  Push a PostgreSQL source tree
to GitHub including this commit, after enabling travis-ci.org and/or
appveyor.com to watch your GitHub repo and build any branch containing
.travis.yml/appveyor.yml.

This is a good way to find out what will happen on cfbot.cputube.org after
you post a patch to the pgsql-hackers mailing list in a thread that is
registered in commitfest.postgresql.org.

Add described at https://wiki.postgresql.org/wiki/Continuous_Integration .